cache loaded n5 blocks for label lookup
Block lists for label ids are stored in a single data block file. When requesting a list of containing blocks for a specific label id, it has to read the entire data block. Instead of ignoring the rest of the loaded data and having to read the data block file again on a subsequent query for a different label id, it now can instead cache the loaded block data. It allows to invalidate all loaded data if needed.